Add files via upload
Mithra is working on a hospital management system, and you need to implement a feature that allows doctors to input their information and generate a doctor's passport. The doctor's passport should include the doctor's name, address, and ward number.



Write a program that takes input from the user for a doctor's name, address, and ward number. Create a class called Doctor object with the provided information and assign it to a Doctor Passport object. Finally, display the doctor's name, address, and ward number from the doctor's passport object.

Input format :
The first line of input consists of a string representing the name of a patient.

The second line of input consists of a string representing the address of a patient.

The third line of input consists of an integer representing the ward number.

Output format :
The output displays the patient's details in separate lines.



Refer to the sample output for formatting specifications.

Code constraints :
In this scenario, the test cases fall under the following constraints:

1 ≤ length of the name, address ≤ 100

1 ≤ ward number ≤ 100

Sample test cases :
Input 1 :
Raju
Chennai
34
Output 1 :
Name: Raju
Address: Chennai
Ward Number: 34
Input 2 :
John Doe
Coimbatore
37
Output 2 :
Name: John Doe
Address: Coimbatore
Ward Number: 37