Add files via upload
Cooper is learning about shapes and geometry. He wants to create a program that helps him calculate the area and perimeter of a rectangle. Implement a Rectangle class that takes the length and breadth as inputs and has attributes to store these values. In the main function, create an object of the Rectangle class, set the length and breadth, and then print out the area and perimeter of the rectangle.

Input format :
The input consists of two space-separated integers representing the length and breadth of a rectangle.

Output format :
The first line of output prints "Area: " followed by the area of the rectangle.

The second line of output prints "Perimeter: " followed by the perimeter of the rectangle.



Refer to the sample output for formatting specifications.

Code constraints :
In this scenario, the test cases fall under the following constraints:

1 ≤ length, breadth ≤ 100

Sample test cases :
Input 1 :
10 20
Output 1 :
Area: 200
Perimeter: 60
Input 2 :
14 18
Output 2 :
Area: 252
Perimeter: 64