Create 01InputAndDisplay
Farhana is fond of coding and she is learning structure and union concepts. While practicing questions she came across a question that used the union concept. Since she is only familiar with structures, she is unable to solve this. 



Help her solve the question. 



The question is given below:

Write a C++ program to store information about an employee using a union. The program should prompt the user to enter the employee ID, name, date of birth (DOB), date of joining (DOJ), and salary. The program should then display the entered information.



Union Name: Employee

Input format :
The first line contains an integer representing the employee ID.

The second line contains a string representing the employee's name.

The third line contains a string representing the date of birth in the format "dd/mm/yyyy".

The fourth line contains a string representing the date of joining in the format "dd/mm/yyyy".

The fifth line contains a floating-point number representing the salary.

Output format :
The first line displays the employee ID of the employee.

The second line displays the employee's name.

The third line displays the date of birth of the employee.

The fourth line displays the employee's date of joining.

The fifth line displays the salary of the employee, rounded off to one decimal place.



Refer to the sample output for formatting specifications.

Code constraints :
The maximum length of the employee's name is 20 characters.

The maximum length of the date of birth and date of joining is 10 characters.

Sample test cases :
Input 1 :
200
Karthick
22/03/1988
09/10/2018
57899.75
Output 1 :
ID : 200
Name : Karthick
DOB : 22/03/1988
DOJ : 09/10/2018
Salary : 57899.8
Input 2 :
202
Jessi
02/12/2000
15/10/2018
28085.31
Output 2 :
ID : 202
Name : Jessi
DOB : 02/12/2000
DOJ : 15/10/2018
Salary : 28085.3