Add files via upload
Sharukh needs to calculate the distance and midpoint of a line segment given its endpoints. Write a program to help Sharukh that reads the coordinates of two endpoints of a line segment, computes the distance between them, and finds the midpoint of the segment. 

Input format :
The first line contains the coordinates of the first endpoint as integers represented as x1 and y1.

The second line contains the coordinates of the second endpoint as integers represented as x2 and y2.

Output format :
The first line prints "Distance: " followed by the distance between the two points, as a float value rounded to two decimal places.

The second line prints"Midpoint: " followed by the midpoint coordinates, as float values rounded to two decimal places.



Refer to the sample output for the exact format.

Code constraints :
In this scenario, the test cases fall under the following constraints:

0 ≤ x1, y1, x2, y2 ≤ 104

Sample test cases :
Input 1 :
1 1
2 2
Output 1 :
Distance: 1.41
Midpoint: 1.50 1.50
Input 2 :
3 2
4 1
Output 2 :
Distance: 1.41
Midpoint: 3.50 1.50