Add files via upload
Vasanth is developing a letter analyzer program. Create a program to classify a single character as a vowel, consonant, or invalid using an enum. Print "vowel" for vowels, "consonant" for consonants, and "Invalid input" for non-alphabetic characters. Ensure the program handles both uppercase and lowercase letters.

Input format :
The input consists of a character.

Output format :
If the input is a vowel, the output displays "vowel".

If the input is a consonant, the output displays "consonant".

Otherwise, the output displays "Invalid input".



Refer to the sample output for formatting specifications.

Code constraints :
The input considers lower-case and upper-case letters as vowels and consonants.

Sample test cases :
Input 1 :
a
Output 1 :
vowel
Input 2 :
Z
Output 2 :
consonant
Input 3 :
1
Output 3 :
Invalid input
Input 4 :
?
Output 4 :
Invalid input