Add files via upload
Kim is creating a calendar application and needs to manage and display the months of the year. Write a program that reads a month number from the user (1 for January, 2 for February, etc.), validates the input, and displays the corresponding month name. If the input is invalid, the program should display an error message.



Requirements:

Create an enum called Month to represent the months of the year.
Use a switch statement to convert the month number to its corresponding month name.
Handle invalid input by displaying "Invalid month input".
Input format :
The input consists of an integer representing the desired month number, n.

Output format :
If the input is a valid month (between 1 and 12), the output displays the corresponding month in uppercase letters.

If the input is invalid, display "Invalid month input".



Refer to the sample output for the formatting specifications.

Code constraints :
n > 0

Sample test cases :
Input 1 :
6
Output 1 :
Month: JUNE
Input 2 :
13
Output 2 :
Invalid month input
Input 3 :
1
Output 3 :
Month: JANUARY