Add files via upload
A hospital wants to store the information of its patients. Help them by providing a code. 



Your code should create a structure patient to store the information of patients coming to the hospital. The information stored will be the patient ID, the patient’s gender (1 for Male and 2 for Female), and the age of the patient. There may be n number of patients visiting the hospital.



Write a program to accept the information for n patients. Write a function to find the total number of male and female patients visited. Also, write a function that prints the patients whose age is greater than 40.



Note: This is a sample question asked in a TCS interview.

Input format :
The first line consists of an integer n, representing the number of patients.

The next n lines contain the patient details, each line consisting of three space-separated integers: ID, gender, and age.

Output format :
The first line displays the count of male and female patients separated by a space.

The second line displays the count of patients above 40 years.



Refer to the sample output for formatting specifications.

Code constraints :
1 ≤ n ≤ 10

Sample test cases :
Input 1 :
3 
110 2 25 
210 2 39
220 1 62 
Output 1 :
1 2
1
Input 2 :
4
115 2 35 
205 2 65 
303 1 52 
403 2 40
Output 2 :
1 3
2