Add files via upload
Harshitha wants to create a program to validate phone numbers entered by users. Write a program that defines a class phone_number with a method to input a phone number and check if it is valid. The phone number should be a 10-digit number. If the entered number is not 10 digits, the program should output an error message. Otherwise, it should display the entered phone number.

Input format :
The input consists of a long integer n, representing the phone number.

Output format :
If the phone number is valid, the program should display the message "Your entered number is: " followed by n.

If the phone number is not valid, the program should display the message "Entered Phone number is wrong".



Refer to the sample output for the format specifications.

Code constraints :
1 ≤ n ≤ 1010

Sample test cases :
Input 1 :
1234567890
Output 1 :
Your entered number is: 1234567890
Input 2 :
12345678
Output 2 :
Entered Phone number is wrong
Input 3 :
645351536b
Output 3 :
Entered Phone number is wrong