Add files via upload
You are developing a module for a text-processing application that will manipulate integers. The module should provide functionalities to set an integer, count the number of digits and find the number of odd digits.



You are tasked with implementing an IntegerManipulator class that performs various operations on a given integer. The class should provide methods to count the number of digits and also find the number of odd digits in the input number.



Use a Non-inline member function.



Note: This is a sample question asked in a mPhasis interview.

Input format :
The input consists of an integer n.

Output format :
The first line displays the number of digits in the number.

The second line displays the number of odd digits in the integer.



Refer to the sample output for formatting specifications.

Code constraints :
1 ≤ n ≤ 109

Sample test cases :
Input 1 :
1124
Output 1 :
Number of digits: 4
Number of odd digits: 2
Input 2 :
2468
Output 2 :
Number of digits: 4
Number of odd digits: 0