Add files via upload
Imagine you are developing a digital diary application that allows users to record important moments and events with timestamps. As part of this application, you need to create a program that fetches the hours, minutes, days, months, and years from the user and stores them in a class member function. The program should then display the date, time, and minutes from another member function in the class.



Your task is to write a program that prompts the user to enter the hours, minutes, days, months, and years. Implement a class that includes a member function to initialize these values and another member function to display the formatted date, time, and minutes. This program will enable users to keep track of their important moments with accurate timestamps in an organized manner within the digital diary application.



Note: This is a sample question asked in a Capgemini interview.

Input format :
The input consists of hours, minutes, days, months, and years in the same order separated by a space.

Output format :
The output displays the date, time, and total minutes as shown in the sample outputs.

If the outputs are not valid, the output displays "Invalid".

Code constraints :
0 < Date <= 31

0 < Month <= 12

0 < Year < 9999

0 < hours < 24

0 < minutes < 60

Sample test cases :
Input 1 :
7 25 6 9 2018
Output 1 :
Date:6-9-2018
Time:7 hrs 25 mins
Total mins:445
Input 2 :
28.6 5 9 3 57
Output 2 :
Invalid