Add files via upload
A university wants to store and analyze the marks of its students. They need a program that allows them to input and retrieve the marks of multiple students. They also want to keep track of the total number of students whose marks have been recorded.



Write a program to create a class named "Student" with the following specifications:



a private data member named "count" that is static and initialized to 0.
a public static member function named "getCount" that returns the value of "count".
a public member function named "setMarks" that sets the marks of a student for a subject
a public member function named "getMarks" that returns the marks of a student for a subject.
a private data member named "marks" that is an array of int with size 5.


Note: This is a sample question asked in an Amcat interview.

Input format :
The first line of the input consists of an integer n, representing the number of students.

The next n lines each contain five space-separated integers, representing the marks of a student in five subjects.

Output format :
The first line contains the total number of students.

The next n lines each contain the marks of the corresponding student in the format "Marks of student X: m1 m2 m3 m4 m5", where X is the student's number (starting from 1) and m1, m2, m3, m4, m5 are the marks of that student.



Refer to the sample output for the formatting specifications.

Code constraints :
1 ≤ n ≤ 10

0 ≤ m1, m2, m3, m4, m5 ≤ 100

Sample test cases :
Input 1 :
3
98 97 91 92 98
98 65 85 37 68
95 65 87 85 89
Output 1 :
Total number of students: 3
Marks of student 1: 98 97 91 92 98 
Marks of student 2: 98 65 85 37 68 
Marks of student 3: 95 65 87 85 89