Add files via upload
An ice cream vendor sells his ice-creams in cone (radius r and height h) and square (side r) shaped containers. However, he is unsure of the quantity that can be filled in the two containers. 



You must write a program that prints the volume of the containers given their respective dimensions as input. 



Your class must be named ‘Icecream’ which has two methods with the same name ‘Quantity’ each having the respective dimensions of the containers as the parameters. 



Method prototypes:

float Quantity(int r) // for square
float Quantity(int r, int h) // for cone


Formulas:

Volume of a square = side * side * side
Volume of a cone = 0.33 * pi * radius * radius * height


Use variable name as M_PI for pi.



Note: This is a sample question asked in a Capgemini interview. 

Input format :
The first line of the input consists of the choice (1 for square, 2 for cone).

If the choice is 1, Enter the side of the square in the next line.

If the choice is 2, Enter the radius and height of the cone separated by a space in the next line.



Note: The input type should be a positive integer.

Output format :
The output displays the volume of the container as a float value rounded off to two decimal places.

Code constraints :
1 ≤ input values ≤ 50

Sample test cases :
Input 1 :
1
4
Output 1 :
64.00
Input 2 :
2
4 5
Output 2 :
82.94