Add files via upload
Rani, a receptionist at Hotel SMS Grand rearranges the flowers in the reception daily. Sidharth, manager of the same Hotel writes a program to display the flower arrangement on the reception TV screen. Today, Rani arranged the flowers as shown below.



﻿



Write a program that takes input from the user, and displays the given pattern, by using the setw manipulator function for spacing.



Note: Make the width length as 2.

Input format :
The input consists of an integer N, representing the number of rows in the pattern.

Output format :
The output prints the pattern of flowers, with each row containing an increasing number of * symbols, formatted with a width of 2 for each symbol.



Refer to the sample output for the formatting specifications.

Code constraints :
1 ≤ N ≤ 10

Sample test cases :
Input 1 :
5
Output 1 :
* 
* * 
* * * 
* * * * 
* * * * * 
Input 2 :
8
Output 2 :
* 
* * 
* * * 
* * * * 
* * * * * 
* * * * * * 
* * * * * * * 
* * * * * * * *