Add files via upload
You have been assigned the responsibility of implementing the Area class for the construction company's software application.



You have declared the following inline functions within the Area class: get(), squareArea(), rectangleArea(), circleArea(), and parallelogramArea(). These functions are used to calculate the area of different geometric shapes.



Write a suitable program to calculate the areas of squares, rectangles, circles, and parallelograms. Provide the complete code solution, including the class definition, method implementations, and the main method where the object is created and the methods are called.



Note: This is a sample question asked in a HCL interview.

Input format :
The first line consists of an integer value representing the side of a square. 

The second line consists of two integers representing the length and breadth of a rectangle, respectively. 

The third line consists of a floating-point value representing the radius of a circle.

The last line consists of two floating-point values representing the base and height of a parallelogram, respectively.

Output format :
The output prints the areas of the square, rectangle, circle, and parallelogram in separate lines with the corresponding labels.

The output should be rounded off to two decimal places for the area of the circle and parallelogram.



Refer to the sample output for formatting specifications.

Code constraints :
Take pi value as 3.14.

Sample test cases :
Input 1 :
4
5 4
6.5
6.2 4.5
Output 1 :
Area of the square: 16
Area of the rectangle: 20
Area of the circle: 132.66
Area of the parallelogram: 27.90
Input 2 :
6
10 12
8
15.2 18.7
Output 2 :
Area of the square: 36
Area of the rectangle: 120
Area of the circle: 200.96
Area of the parallelogram: 284.24